Casi terminado (tiene errores)

+++ tp2/estudiantes/src/main/java/aed/Estudiante.java
@@ -3,7 +3,7 @@ package aed;
 public class Estudiante {
 
     private int _id;
-    private int _puntaje;
+    private double _puntaje;
     private boolean _entrego;
     private int[] _respuestas;
     private int _cantRespuestasCorrectas;
@@ -20,7 +20,7 @@ public class Estudiante {
         }
     }
 
-    public int puntaje() {return _puntaje;}
+    public double puntaje() {return _puntaje;}
     public boolean entrego() {return _entrego;}
     public int[] respuestas() {return _respuestas;}
     public int id() {return _id;}
@@ -30,7 +30,7 @@ public class Estudiante {
         
         if(examenCanonico[ejercicio] == _respuestas[ejercicio]){
             _cantRespuestasCorrectas ++;
-            _puntaje = (int) Math.floor(_cantRespuestasCorrectas * 100 / examenCanonico.length);
+            _puntaje = Math.floor(_cantRespuestasCorrectas * 100 / examenCanonico.length);
         }
     }
 
@@ -44,7 +44,7 @@ public class Estudiante {
             }
         }
         _cantRespuestasCorrectas = cantRespuestasCorrectas;
-        _puntaje = (int) Math.floor(cantRespuestasCorrectas * 100 / examenCanonico.length);
+        _puntaje = Math.floor(cantRespuestasCorrectas * 100 / examenCanonico.length);
     }
 
     public void entregar(){_entrego = true;}
